+++ libs/ui/u-auth/src/lib/auth-routing.module.ts
@@ -7,6 +7,7 @@ import { ResetPasswordFormContainerComponent } from './containers/reset-password
 import { SetPasswordFormContainerComponent } from './containers/set-password-form-container/set-password-form-container.component';
 import { SignInFormContainerComponent } from './containers/sign-in-form-container/sign-in-form-container.component';
 import { SetPasswordGuard } from './guards/set-password.guard';
+import { SpaGuard } from './guards/spa.guard';
 import { SetPasswordResolverService } from './resolvers/set-password-resolver.service';
 
 export const routes: Routes = [
@@ -17,10 +18,12 @@ export const routes: Routes = [
       {
         path: 'sign-in',
         component: SignInFormContainerComponent,
+        canActivate: [SpaGuard],
       },
       {
         path: 'reset-password',
         component: ResetPasswordFormContainerComponent,
+        canActivate: [SpaGuard],
       },
       {
         path: 'set-password/:passwordToken/:userId',