+++ libs/ui/u-auth/src/lib/guards/auth.guard.ts
@@ -1,9 +1,7 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, Router } from '@angular/router';
-import { AuthService, fromAuth } from '@ivt/u-state';
-import { Actions, ofType } from '@ngrx/effects';
-import { of } from 'rxjs';
-import { map, switchMap, take } from 'rxjs/operators';
+import { AuthService } from '@ivt/u-state';
+import { map, take } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root',
@@ -11,20 +9,12 @@ import { map, switchMap, take } from 'rxjs/operators';
 export class AuthGuard implements CanActivate {
   loginUrlTree = this.router.parseUrl('/auth');
 
-  constructor(private router: Router, private authService: AuthService, private actions$: Actions) {}
+  constructor(private router: Router, private authService: AuthService) {}
 
   canActivate() {
     return this.authService.isAuthenticated().pipe(
       take(1),
-      switchMap(isAuthenticated =>
-        isAuthenticated
-          ? of(true)
-          : this.actions$.pipe(
-              ofType(fromAuth.actions.loadUserFromStorage),
-              take(1),
-              map(({ user }) => (user.token ? true : this.loginUrlTree))
-            )
-      )
+      map(isAuthenticated => (isAuthenticated ? true : this.loginUrlTree))
     );
   }
 }